break down github widget in smaller pieces
