+++ src/gtnh/gui/github/github_modlist.py
@@ -1,12 +1,10 @@
 import asyncio
 from tkinter import END, Button, Entry, Label, LabelFrame, Listbox, Scrollbar, StringVar
 from tkinter.messagebox import showerror, showinfo, showwarning
-from tkinter.ttk import Combobox
 from typing import Any, Callable, Coroutine, Dict, List, Optional
 
 from gtnh.defs import Position
 from gtnh.exceptions import RepoNotFoundException
-from gtnh.gui.mod_info_frame import ModInfoFrame
 from gtnh.models.gtnh_version import GTNHVersion
 from gtnh.models.mod_info import GTNHModInfo
 from gtnh.modpack_manager import GTNHModpackManager
@@ -324,303 +322,3 @@ class GithubModList(LabelFrame):
         await gtnh.regen_github_assets(callback=self._update_callback)
         showinfo("Github assets had been updated successfully", "All the github assets had been updated successfully!")
 
-
-class GithubModFrame(LabelFrame):
-    """
-    Main frame widget for the github mods' management.
-    """
-
-    def __init__(
-        self,
-        master: Any,
-        frame_name: str,
-        callbacks: Dict[str, Any],
-        width: Optional[int] = None,
-        **kwargs: Any,
-    ):
-        """
-        Constructor of the GithubModFrame class.
-
-        :param master: the parent widget
-        :param frame_name: the name displayed in the framebox
-        :param callbacks: a dict of callbacks passed to this instance
-        :param width: the width to harmonize widgets in characters
-        :param kwargs: params to init the parent class
-        """
-        LabelFrame.__init__(self, master, text=frame_name, **kwargs)
-        self.ypadding: int = 0
-        self.xpadding: int = 0
-        self.width: Optional[int] = width
-
-        modpack_version_callbacks: Dict[str, Any] = {
-            "set_modpack_version": callbacks["set_modpack_version"],
-            "get_gtnh": callbacks["get_gtnh"],
-        }
-
-        self.modpack_version_frame: ModpackVersionFrame = ModpackVersionFrame(
-            self, frame_name="Modpack version", callbacks=modpack_version_callbacks
-        )
-
-        mod_info_callbacks: Dict[str, Any] = {
-            "set_mod_version": callbacks["set_github_mod_version"],
-            "set_mod_side": callbacks["set_github_mod_side"],
-        }
-
-        self.mod_info_frame: ModInfoFrame = ModInfoFrame(
-            self, frame_name="Github mod info", callbacks=mod_info_callbacks
-        )
-
-        github_mod_list_callbacks: Dict[str, Any] = {
-            "mod_info": self.mod_info_frame.populate_data,
-            "get_github_mods": callbacks["get_github_mods"],
-            "get_gtnh": callbacks["get_gtnh"],
-            "reset_mod_info": self.mod_info_frame.reset,
-            "update_current_task_progress_bar": callbacks["update_current_task_progress_bar"],
-            "update_global_progress_bar": callbacks["update_global_progress_bar"],
-            "reset_current_task_progress_bar": callbacks["reset_current_task_progress_bar"],
-            "reset_global_progress_bar": callbacks["reset_global_progress_bar"],
-            "add_mod_in_memory": callbacks["add_mod_in_memory"],
-            "del_mod_in_memory": callbacks["del_mod_in_memory"],
-        }
-
-        self.github_mod_list: GithubModList = GithubModList(
-            self, frame_name="Github mod list", callbacks=github_mod_list_callbacks
-        )
-
-        if self.width is None:
-            self.width = self.github_mod_list.get_width()
-        else:
-            self.github_mod_list.set_width(self.width)
-
-        self.mod_info_frame.set_width(self.width)
-        self.modpack_version_frame.set_width(self.width)
-
-        self.update_widget()
-
-    def configure_widgets(self) -> None:
-        """
-        Method to configure the widgets.
-
-        :return: None
-        """
-        self.mod_info_frame.configure_widgets()
-        self.modpack_version_frame.configure_widgets()
-        self.github_mod_list.configure_widgets()
-
-    def set_width(self, width: int) -> None:
-        """
-        Method to set the widgets' width.
-
-        :param width: the new width
-        :return: None
-        """
-        self.width = width
-        self.mod_info_frame.set_width(self.width)
-        self.modpack_version_frame.set_width(self.width)
-        self.github_mod_list.set_width(self.width)
-
-    def get_width(self) -> int:
-        """
-        Getter for self.width.
-
-        :return: the width in character sizes of the normalised widgets
-        """
-        assert self.width  # can't be None because how it's defined in the constructor
-        return self.width
-
-    def update_widget(self) -> None:
-        """
-        Method to update the widget and all its childs
-
-        :return: None
-        """
-        self.hide()
-        self.configure_widgets()
-        self.show()
-
-        self.modpack_version_frame.update_widget()
-        self.mod_info_frame.update_widget()
-        self.github_mod_list.update_widget()
-
-    def hide(self) -> None:
-        """
-        Method to hide the widget and all its childs
-        :return None:
-        """
-        self.modpack_version_frame.grid_forget()
-        self.github_mod_list.grid_forget()  # ref widget
-        self.mod_info_frame.grid_forget()
-
-        self.modpack_version_frame.hide()
-        self.github_mod_list.hide()
-        self.mod_info_frame.hide()
-
-        self.update_idletasks()
-
-    def show(self) -> None:
-        """
-        Method used to display widgets and child widgets, as well as to configure the "responsiveness" of the widgets.
-
-        :return: None
-        """
-        x: int = 0
-        y: int = 0
-        rows: int = 3
-        columns: int = 1
-
-        for i in range(rows):
-            self.rowconfigure(i, weight=1, pad=self.xpadding)
-
-        for i in range(columns):
-            self.columnconfigure(i, weight=1, pad=self.ypadding)
-
-        self.modpack_version_frame.grid(row=x, column=y, sticky=Position.HORIZONTAL)
-        self.github_mod_list.grid(row=x + 1, column=y)  # ref widget
-        self.mod_info_frame.grid(row=x + 2, column=y, sticky=Position.HORIZONTAL)
-
-        self.modpack_version_frame.show()
-        self.github_mod_list.show()
-        self.mod_info_frame.show()
-
-        self.update_idletasks()
-
-    def populate_data(self, data: Dict[str, Any]) -> None:
-        """
-        Method called by parent class to populate data in this class.
-
-        :param data: the data to pass to this class
-        :return: None
-        """
-        self.github_mod_list.populate_data(data["github_mod_list"])
-        self.modpack_version_frame.populate_data(data["modpack_version_frame"])
-
-
-class ModpackVersionFrame(LabelFrame):
-    """
-    Frame to chose the gtnh modpack repo assets' version.
-    """
-
-    def __init__(
-        self,
-        master: Any,
-        frame_name: str,
-        callbacks: Dict[str, Any],
-        width: Optional[int] = None,
-        **kwargs: Any,
-    ):
-        """
-        Constructor of the ModpackVersionFrame.
-
-        :param master: the parent widget
-        :param frame_name: the name displayed in the framebox
-        :param callbacks: a dict of callbacks passed to this instance
-        :param width: the width to harmonize widgets in characters
-        :param kwargs: params to init the parent class
-        """
-        LabelFrame.__init__(self, master, text=frame_name, **kwargs)
-        self.ypadding: int = 0
-        self.xpadding: int = 0
-        modpack_version_text: str = "Modpack version:"
-        refresh_modpack_text: str = "Refresh modpack assets"
-        self.width: int = width if width is not None else max(len(modpack_version_text), len(refresh_modpack_text))
-        self.label_modpack_version: Label = Label(self, text=modpack_version_text)
-        self.sv_version: StringVar = StringVar(value="")
-        self.cb_modpack_version: Combobox = Combobox(self, textvariable=self.sv_version, values=[])
-        self.cb_modpack_version.bind(
-            "<<ComboboxSelected>>", lambda event: callbacks["set_modpack_version"](self.sv_version.get())
-        )
-        self.btn_refresh: Button = Button(
-            self, text=refresh_modpack_text, command=lambda: asyncio.ensure_future(self.refresh_modpack_assets())
-        )
-        self.get_gtnh_callback: Callable[[], Coroutine[Any, Any, GTNHModpackManager]] = callbacks["get_gtnh"]
-
-    async def refresh_modpack_assets(self) -> None:
-        """
-        Method used to refresh assets for the modpack repository.
-
-        :return: None
-        """
-        gtnh: GTNHModpackManager = await self.get_gtnh_callback()
-        await gtnh.regen_config_assets()
-        self.cb_modpack_version["values"] = [version.version_tag for version in gtnh.assets.config.versions]
-        self.sv_version.set(gtnh.assets.config.latest_version)
-        showinfo("Modpack assets refreshed", "Modpack assets refreshed successfully!")
-
-    def configure_widgets(self) -> None:
-        """
-        Method to configure the widgets.
-
-        :return: None
-        """
-        self.label_modpack_version.configure(width=self.width)
-        self.cb_modpack_version.configure(width=self.width)
-
-    def set_width(self, width: int) -> None:
-        """
-        Method to set the widgets' width.
-
-        :param width: the new width
-        :return: None
-        """
-        self.width = width
-        self.configure_widgets()
-
-    def get_width(self) -> int:
-        """
-        Getter for self.width.
-
-        :return: the width in character sizes of the normalised widgets
-        """
-        return self.width
-
-    def update_widget(self) -> None:
-        """
-        Method to update the widget and all its childs
-
-        :return: None
-        """
-        self.hide()
-        self.configure_widgets()
-        self.show()
-
-    def hide(self) -> None:
-        """
-        Method to hide the widget and all its childs
-        :return None:
-        """
-        self.label_modpack_version.grid_forget()
-        self.cb_modpack_version.grid_forget()
-        self.btn_refresh.grid_forget()
-
-        self.update_idletasks()
-
-    def show(self) -> None:
-        """
-        Method used to display widgets and child widgets, as well as to configure the "responsiveness" of the widgets.
-
-        :return: None
-        """
-        x: int = 0
-        y: int = 0
-        rows: int = 2
-        columns: int = 2
-
-        for i in range(rows):
-            self.rowconfigure(i, weight=1, pad=self.xpadding)
-
-        for i in range(columns):
-            self.columnconfigure(i, weight=1, pad=self.ypadding)
-
-        self.label_modpack_version.grid(row=x, column=y)
-        self.cb_modpack_version.grid(row=x, column=y + 1, sticky=Position.HORIZONTAL)
-        self.btn_refresh.grid(row=x + 1, column=y + 1, sticky=Position.HORIZONTAL)
-
-    def populate_data(self, data: Dict[str, Any]) -> None:
-        """
-        Method called by parent class to populate data in this class.
-
-        :param data: the data to pass to this class
-        :return: None
-        """
-        self.cb_modpack_version["values"] = data["combobox"]
-        self.sv_version.set(data["stringvar"])
